Migrate the class feature sheet to AppV2
